worker/requirements.txt
# Worker dependencies

# Database (shared with API)
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis and Queue
redis==5.0.1
hiredis==2.3.2
rq==1.15.1

# LLM Providers
google-generativeai==0.3.2
openai==1.10.0
anthropic==0.8.1

# LangChain
langchain>=0.1.6,<0.2
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-anthropic

# Text Processing
langchain-text-splitters
tiktoken==0.5.2

# Document Processing
python-docx==1.1.0
pypdf==4.0.1
markdown-it-py==3.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Object Storage (S3-compatible)
boto3==1.34.36
botocore==1.34.36

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0

# Logging
structlog==24.1.0

# Utilities
tenacity==8.2.3  # Retry logic
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Metrics and monitoring
prometheus-client==0.19.0

# JSON handling
orjson==3.9.13

# Error Tracking
sentry-sdk==1.40.0  # Error tracking and performance monitoring
